replaced float with int to create mask

+++ load.py
@@ -55,7 +55,7 @@ def sample(df, n, shape):
     for image_file, label_file in zip(image_files, label_files):
         image = load_nifti(image_file)
         label = load_nifti(label_file).astype(np.int32)
-        mask = np.float32(label > 0)
+        mask = np.int32(label > 0)
         slices = [slice(len_ / 2, -len_ / 2) for len_ in shape]
         mask[slices] *= 2
         indices = np.where(mask > 1.5)
